% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanshiny.R
\name{lavaan.shiny}
\alias{lavaan.shiny}
\title{Launch lavaan.shiny User Interface}
\usage{
lavaan.shiny()
}
\value{
Nothing
}
\description{
lavaan.shiny() loads interactive user interface built using R shiny.
}
\details{
Start lavaan.shiny

The interactive user interface is to provide an easy way for people who are learning how to work with the lavaan package and/or are not comfortable with the R command line system. Includes example data for testing out a few example analysis.
}
\examples{
\dontrun{
library(shiny)
lavaan.shiny()
}
}
\keyword{lavaan.shiny}

